+++ jni/test/jarray_test.dart
@@ -21,7 +21,7 @@ void main() {
 void run({required TestRunnerCallback testRunner}) {
   testRunner("Java boolean array", () {
     using((arena) {
-      final array = JArray(JBoolean.type, 3)..deletedIn(arena);
+      final array = JArray(jboolean.type, 3)..deletedIn(arena);
       expect(array.length, 3);
       array[0] = true;
       array[1] = false;
@@ -46,7 +46,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner("Java char array", () {
     using((arena) {
-      final array = JArray(JChar.type, 3)..deletedIn(arena);
+      final array = JArray(jchar.type, 3)..deletedIn(arena);
       expect(array.length, 3);
       array[0] = 'ح';
       array[1] = '2';
@@ -71,7 +71,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner("Java byte array", () {
     using((arena) {
-      final array = JArray(JByte.type, 3)..deletedIn(arena);
+      final array = JArray(jbyte.type, 3)..deletedIn(arena);
       expect(array.length, 3);
       array[0] = 1;
       array[1] = 2;
@@ -96,7 +96,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner("Java short array", () {
     using((arena) {
-      final array = JArray(JShort.type, 3)..deletedIn(arena);
+      final array = JArray(jshort.type, 3)..deletedIn(arena);
       expect(array.length, 3);
       array[0] = 1;
       array[1] = 2;
@@ -121,7 +121,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner("Java int array", () {
     using((arena) {
-      final array = JArray(JInt.type, 3)..deletedIn(arena);
+      final array = JArray(jint.type, 3)..deletedIn(arena);
       expect(array.length, 3);
       array[0] = 1;
       array[1] = 2;
@@ -147,7 +147,7 @@ void run({required TestRunnerCallback testRunner}) {
   const epsilon = 1e-6;
   testRunner("Java float array", () {
     using((arena) {
-      final array = JArray(JFloat.type, 3)..deletedIn(arena);
+      final array = JArray(jfloat.type, 3)..deletedIn(arena);
       expect(array.length, 3);
       array[0] = 0.5;
       array[1] = 2;
@@ -172,7 +172,7 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner("Java double array", () {
     using((arena) {
-      final array = JArray(JDouble.type, 3)..deletedIn(arena);
+      final array = JArray(jdouble.type, 3)..deletedIn(arena);
       expect(array.length, 3);
       array[0] = 0.5;
       array[1] = 2;
@@ -241,11 +241,11 @@ void run({required TestRunnerCallback testRunner}) {
   });
   testRunner("Java 2d array", () {
     using((arena) {
-      final array = JArray(JInt.type, 3)..deletedIn(arena);
+      final array = JArray(jint.type, 3)..deletedIn(arena);
       array[0] = 1;
       array[1] = 2;
       array[2] = 3;
-      final twoDimArray = JArray(JArray.type(JInt.type), 3)..deletedIn(arena);
+      final twoDimArray = JArray(JArray.type(jint.type), 3)..deletedIn(arena);
       expect(twoDimArray.length, 3);
       twoDimArray[0] = array;
       twoDimArray[1] = array;